requirements.txt
asgiref==3.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.13.1
djongo==1.3.6
mongoengine==0.24.1
Pillow==9.1.0
pymongo==3.12.3
python-decouple==3.6
pytz==2022.1
sqlparse==0.2.4
